requirements.txt
fastapi==0.115.0
uvicorn==0.30.0

# Agent Orchestration
langgraph==0.2.28
langchain-core==0.2.43

# Google Vertex AI (Gemini)
google-cloud-aiplatform==1.60.0

# Config & Env
pydantic-settings==2.3.0
pydantic==2.8.0
python-dotenv==1.0.0

# Vector DB (RAG)
chromadb==0.4.24
posthog==3.5.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# File handling
python-multipart==0.0.9
aiofiles==23.2.1

# Database
sqlalchemy==2.0.36